// src/types/form-values.ts
export interface IFormValues {
  annualAppreciationRate: number;
  annualInsurance: number;
  initPropertyValue: number;
  interestRate: number;
  landValue: number;
  maintenanceRate: number;
  monthlyRentalIncome: number;
  percentDown: number;
  propertyManagementRate: number;
  propertyTaxRate: number;
  totalYears: number;
}